Moving things from enc -> encrypt
